Add createRecord tests
