+++ sdk-core/src/test/java/care/data4life/sdk/RecordServiceCreateRecordTest.kt
@@ -15,317 +15,372 @@
  */
 package care.data4life.sdk
 
-/*
-class RecordServiceCreateRecordTest : RecordServiceTestBase() {
+
+import care.data4life.crypto.GCKey
+import care.data4life.sdk.call.DataRecord
+import care.data4life.sdk.call.Fhir4Record
+import care.data4life.sdk.data.DataResource
+import care.data4life.sdk.fhir.Fhir3Resource
+import care.data4life.sdk.fhir.Fhir4Resource
+import care.data4life.sdk.model.ModelVersion
+import care.data4life.sdk.model.Record
+import care.data4life.sdk.model.SdkRecordFactory
+import care.data4life.sdk.model.definitions.BaseRecord
+import care.data4life.sdk.network.DecryptedRecordBuilder
+import care.data4life.sdk.network.model.EncryptedRecord
+import care.data4life.sdk.network.model.NetworkRecordContract
+import care.data4life.sdk.wrapper.ResourceFactory
+import care.data4life.sdk.wrapper.ResourceHelper
+import care.data4life.sdk.wrapper.WrapperContract
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.mockkConstructor
+import io.mockk.mockkObject
+import io.mockk.unmockkConstructor
+import io.mockk.unmockkObject
+import io.mockk.verify
+import io.reactivex.Single
+import org.junit.After
+import org.junit.Assert.assertSame
+import org.junit.Before
+import org.junit.Ignore
+import org.junit.Test
+import kotlin.collections.HashMap
+
+
+class RecordServiceCreateRecordTest: RecordTestBase() {
+    private lateinit var builder: NetworkRecordContract.Builder
+
     @Before
-    fun setup() {
+    fun setUp() {
         init()
+
+        builder = mockk()
+
+        mockkObject(ResourceFactory)
+        mockkObject(ResourceHelper)
+        mockkObject(SdkRecordFactory)
+        mockkConstructor(DecryptedRecordBuilder::class)
     }
 
     @After
     fun tearDown() {
-        stop()
+        unmockkObject(ResourceFactory)
+        unmockkObject(ResourceHelper)
+        unmockkObject(SdkRecordFactory)
+        unmockkConstructor(DecryptedRecordBuilder::class)
     }
 
     @Test
-    @Throws(
-            InterruptedException::class,
-            IOException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class,
-            DataValidationException.ModelVersionNotSupported::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class
-    )
-    fun `Given, createRecord is called with a DomainResource and a UserId, it returns a new Record`() {
+    fun `Given, createRecord is called with a DataResource, it wraps it and delegates it to the generic createRecord and return its Result`() {
         // Given
+        val userId = "id"
+        val rawResource = mockk<DataResource>()
+
+        val wrappedResource =  mockk<WrapperContract.Resource>()
+
+        val createdRecord = mockk<DataRecord<DataResource>>()
+
+        every { ResourceFactory.wrap(rawResource) } returns wrappedResource
+
+        @Suppress("UNCHECKED_CAST")
         every {
-            anyConstructed<DecryptedRecordBuilderImpl>().setAnnotations(listOf())
-        } returns mockDecryptedRecordBuilder as DecryptedRecordBuilderImpl
-
-        @Suppress("UNCHECKED_CAST")
-        Mockito.`when`(
-                mockDecryptedRecordBuilder.build(
-                        mockCarePlan,
-                        mockTags,
-                        DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                        mockDataKey,
-                        ModelVersion.CURRENT
-                )
-        ).thenReturn(mockDecryptedFhir3Record as DecryptedFhir3Record<CarePlan>)
-        Mockito.doReturn(mockUploadData).`when`(recordService).extractUploadData(mockCarePlan)
-        Mockito.`when`(mockCarePlan.resourceType).thenReturn(CarePlan.resourceType)
-        Mockito.`when`(mockTaggingService.appendDefaultTags(
-                ArgumentMatchers.eq(CarePlan.resourceType),
-                ArgumentMatchers.any<HashMap<String, String>>())
-        ).thenReturn(mockTags)
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockDataKey))
-        Mockito.doReturn(mockDecryptedFhir3Record).`when`(recordService)
-                ._uploadData(
-                        mockDecryptedFhir3Record,
-                        USER_ID
-                )
-        Mockito.doReturn(mockDecryptedFhir3Record).`when`(recordService)
-                .removeUploadData(mockDecryptedFhir3Record)
-
-        Mockito.doReturn(mockEncryptedRecord)
-                .`when`(recordService).encryptRecord(
-                        mockDecryptedFhir3Record
-                )
-        Mockito.`when`(mockApiService.createRecord(ALIAS, USER_ID, mockEncryptedRecord))
-                .thenReturn(Single.just(mockEncryptedRecord))
-        Mockito.doReturn(mockDecryptedFhir3Record)
-                .`when`(recordService)
-                .decryptRecord<DomainResource>(mockEncryptedRecord, USER_ID)
-        Mockito.doReturn(mockDecryptedFhir3Record)
-                .`when`(recordService).restoreUploadData(
-                        mockDecryptedFhir3Record,
-                        mockCarePlan,
-                        mockUploadData
-                )
-        @Suppress("UNCHECKED_CAST")
-        every { SdkRecordFactory.getInstance(mockDecryptedFhir3Record) } returns mockRecord as BaseRecord<DomainResource>
-
-        val annotations = listOf<String>()
+            recordService.createRecord(userId, wrappedResource, any())
+        } returns Single.just(createdRecord as BaseRecord<Any>)
 
         // When
-        val subscriber = recordService.createRecord(USER_ID, mockCarePlan, annotations).test().await()
+        val subscriber = recordService.createRecord(
+                userId,
+                rawResource,
+                mockk(relaxed = true)
+        ).test().await()
 
-        // Then
         val record = subscriber
                 .assertNoErrors()
                 .assertComplete()
                 .assertValueCount(1)
                 .values()[0]
-        Truth.assertThat(record).isSameInstanceAs(mockRecord)
 
-        inOrder.verify(mockTaggingService).appendDefaultTags(
-                ArgumentMatchers.anyString(),
-                ArgumentMatchers.any<HashMap<String, String>>()
+        // Then
+        assertSame(
+                record,
+                createdRecord
         )
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockDecryptedRecordBuilder).build(
-                mockCarePlan,
-                mockTags,
-                DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                mockDataKey,
-                ModelVersion.CURRENT
-        )
-        inOrder.verify(recordService)._uploadData(
-                mockDecryptedFhir3Record,
-                USER_ID
-        )
-        inOrder.verify(recordService).removeUploadData(mockDecryptedFhir3Record)
 
-        inOrder.verify(recordService).encryptRecord(
-                mockDecryptedFhir3Record
-        )
-        inOrder.verify(mockApiService).createRecord(ALIAS, USER_ID, mockEncryptedRecord)
-        inOrder.verify(recordService).decryptRecord<DomainResource>(mockEncryptedRecord, USER_ID)
-        inOrder.verify(recordService).restoreUploadData(
-                mockDecryptedFhir3Record,
-                mockCarePlan,
-                mockUploadData
-        )
-        inOrder.verify(recordService).assignResourceId(mockDecryptedFhir3Record)
-        inOrder.verifyNoMoreInteractions()
+        verify(exactly = 1) { ResourceFactory.wrap(rawResource) }
     }
 
     @Test
-    @Throws(
-            InterruptedException::class,
-            IOException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class,
-            DataValidationException.ModelVersionNotSupported::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class
-    )
-    fun `Given, createRecord is called a DomainResource, a UserId and no attachment, it returns a new Record`() {
+    fun `Given, createRecord is called with a Fhir3Resource, it wraps it and delegates it to the generic createRecord and return its Result`() {
         // Given
+        val userId = "id"
+        val rawResource = mockk<Fhir3Resource>()
+
+        val wrappedResource =  mockk<WrapperContract.Resource>()
+
+        val createdRecord = mockk<Record<Fhir3Resource>>()
+
+        every { ResourceFactory.wrap(rawResource) } returns wrappedResource
+
+        @Suppress("UNCHECKED_CAST")
         every {
-            anyConstructed<DecryptedRecordBuilderImpl>().setAnnotations(listOf())
-        } returns mockDecryptedRecordBuilder as DecryptedRecordBuilderImpl
-
-        @Suppress("UNCHECKED_CAST")
-        Mockito.`when`(
-                mockDecryptedRecordBuilder.build(
-                        mockCarePlan,
-                        mockTags,
-                        DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                        mockDataKey,
-                        ModelVersion.CURRENT
-                )
-        ).thenReturn(mockDecryptedFhir3Record as DecryptedFhir3Record<CarePlan>)
-        Mockito.doReturn(null).`when`(recordService).extractUploadData(mockCarePlan)
-        Mockito.`when`(mockCarePlan.resourceType).thenReturn(CarePlan.resourceType)
-        Mockito.`when`(
-                mockTaggingService.appendDefaultTags(
-                        ArgumentMatchers.eq(CarePlan.resourceType),
-                        ArgumentMatchers.any<HashMap<String, String>>()
-                )
-        ).thenReturn(mockTags)
-        Mockito.`when`(mockCryptoService.generateGCKey())
-                .thenReturn(Single.just(mockDataKey))
-        Mockito.doReturn(mockDecryptedFhir3Record)
-                .`when`(recordService)
-                ._uploadData(
-                        mockDecryptedFhir3Record,
-                        USER_ID
-                )
-        Mockito.doReturn(mockDecryptedFhir3Record)
-                .`when`(recordService)
-                .removeUploadData(mockDecryptedFhir3Record)
-
-        Mockito.doReturn(mockEncryptedRecord).`when`(recordService)
-                .encryptRecord(mockDecryptedFhir3Record)
-        Mockito.`when`(mockApiService.createRecord(ALIAS, USER_ID, mockEncryptedRecord))
-                .thenReturn(Single.just(mockEncryptedRecord))
-        Mockito.doReturn(mockDecryptedFhir3Record)
-                .`when`(recordService)
-                .decryptRecord<DomainResource>(
-                        mockEncryptedRecord,
-                        USER_ID
-                )
-        Mockito.doReturn(mockDecryptedFhir3Record)
-                .`when`(recordService)
-                .restoreUploadData(
-                        mockDecryptedFhir3Record,
-                        mockCarePlan,
-                        mockUploadData
-                )
-
-        Mockito.doReturn(mockDecryptedFhir3Record)
-                .`when`(recordService)
-                .uploadData(
-                        mockDecryptedFhir3Record,
-                        null,
-                        USER_ID
-                )
-        @Suppress("UNCHECKED_CAST")
-        every { SdkRecordFactory.getInstance(mockDecryptedFhir3Record) } returns mockRecord as BaseRecord<DomainResource>
-
-        val annotations = listOf<String>()
+            recordService.createRecord(userId, wrappedResource, any())
+        } returns Single.just(createdRecord as BaseRecord<Any>)
 
         // When
-        val subscriber = recordService.createRecord(USER_ID, mockCarePlan, annotations).test().await()
+        val subscriber = recordService.createRecord(
+                userId,
+                rawResource,
+                mockk(relaxed = true)
+        ).test().await()
 
-        // Then
         val record = subscriber
                 .assertNoErrors()
                 .assertComplete()
                 .assertValueCount(1)
                 .values()[0]
-        Truth.assertThat(record).isSameInstanceAs(mockRecord)
 
-        inOrder.verify(mockTaggingService).appendDefaultTags(
-                ArgumentMatchers.anyString(),
-                ArgumentMatchers.any<HashMap<String, String>>()
+        // Then
+        assertSame(
+                record,
+                createdRecord
         )
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockDecryptedRecordBuilder).build(
-                mockCarePlan,
-                mockTags,
-                DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                mockDataKey,
+
+        verify(exactly = 1) { ResourceFactory.wrap(rawResource) }
+    }
+
+    @Test
+    fun `Given, createRecord is called with a Fhir4Resource, it wraps it and delegates it to the generic createRecord and return its Result`() {
+        // Given
+        val userId = "id"
+        val rawResource = mockk<Fhir4Resource>()
+
+        val wrappedResource =  mockk<WrapperContract.Resource>()
+
+        val createdRecord = mockk<Fhir4Record<Fhir4Resource>>()
+
+        every { ResourceFactory.wrap(rawResource) } returns wrappedResource
+
+        @Suppress("UNCHECKED_CAST")
+        every {
+            recordService.createRecord(userId, wrappedResource, any())
+        } returns Single.just(createdRecord as BaseRecord<Any>)
+
+        // When
+        val subscriber = recordService.createRecord(
+                userId,
+                rawResource,
+                mockk(relaxed = true)
+        ).test().await()
+
+        val record = subscriber
+                .assertNoErrors()
+                .assertComplete()
+                .assertValueCount(1)
+                .values()[0]
+
+        // Then
+        assertSame(
+                record,
+                createdRecord
+        )
+
+        verify(exactly = 1) { ResourceFactory.wrap(rawResource) }
+    }
+
+    @Test
+    fun `Given, createRecord is called with a wrapped Resource, it return a BaseRecord`() {
+        // Given
+        val userId = "id"
+        val rawResource = mockk<Fhir4Resource>()
+
+        val wrappedResource =  mockk<WrapperContract.Resource>()
+
+        val createdRecord = mockk<Fhir4Record<Fhir4Resource>>()
+
+        every { ResourceFactory.wrap(rawResource) } returns wrappedResource
+
+        @Suppress("UNCHECKED_CAST")
+        every {
+            recordService.createRecord(userId, wrappedResource, any())
+        } returns Single.just(createdRecord as BaseRecord<Any>)
+
+        // When
+        val subscriber = recordService.createRecord(
+                userId,
+                rawResource,
+                mockk(relaxed = true)
+        ).test().await()
+
+        val record = subscriber
+                .assertNoErrors()
+                .assertComplete()
+                .assertValueCount(1)
+                .values()[0]
+
+        // Then
+        assertSame(
+                record,
+                createdRecord
+        )
+
+        verify(exactly = 1) { ResourceFactory.wrap(rawResource) }
+    }
+
+
+    @Test
+    fun `Given, createRecord is called with a DataResource and a UserId, it returns a new Record`() {
+        // Given
+        val wrappedResource =  mockk<WrapperContract.Resource>()
+        val annotations = mockk<List<String>>()
+        val userId = "id"
+
+        val data = mockk<HashMap<WrapperContract.Attachment, String?>>()
+        val decryptedRecord = mockk<NetworkRecordContract.DecryptedRecord>()
+        val tags = mockk<HashMap<String, String>>()
+        val dataKey = mockk<GCKey>()
+
+        val encryptedRecord = mockk<EncryptedRecord>()
+
+        val createdRecord = mockk<BaseRecord<Any>>()
+
+
+        every { wrappedResource.type } returns WrapperContract.Resource.TYPE.DATA
+
+        every { attachmentService.checkDataRestrictions(wrappedResource) } returns mockk()
+        every { attachmentService.extractUploadData(wrappedResource) } returns data
+
+        every { taggingService.appendDefaultTags( null, null ) } returns tags
+
+        every { cryptoService.generateGCKey() } returns Single.just(dataKey)
+
+        every {
+            anyConstructed<DecryptedRecordBuilder>().setAnnotations(annotations)
+        } returns builder
+        every { builder.build(
+                wrappedResource,
+                tags,
+                any(),//FIXME
+                dataKey,
                 ModelVersion.CURRENT
-        )
-        inOrder.verify(recordService)._uploadData(
-                mockDecryptedFhir3Record,
-                USER_ID
-        )
-        inOrder.verify(recordService).removeUploadData(mockDecryptedFhir3Record)
-        inOrder.verify(recordService).encryptRecord(mockDecryptedFhir3Record)
-        inOrder.verify(mockApiService).createRecord(ALIAS, USER_ID, mockEncryptedRecord)
-        inOrder.verify(recordService).decryptRecord<DomainResource>(mockEncryptedRecord, USER_ID)
-        inOrder.verify(recordService).restoreUploadData(
-                mockDecryptedFhir3Record,
-                mockCarePlan,
-                null
-        )
-        inOrder.verify(recordService).assignResourceId(mockDecryptedFhir3Record)
-        inOrder.verifyNoMoreInteractions()
-    }
+        ) } returns decryptedRecord
 
-    @Test
-    @Throws(
-            InterruptedException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class
-    )
-    fun `Given, createRecord is called with unsupported data, it throws an error`() {
-        // Given
-        val invalidData = byteArrayOf(0x00)
-        val doc = buildDocumentReference(invalidData)
-        val annotations = listOf<String>()
+        every { attachmentClient.uploadData(decryptedRecord, userId) } returns decryptedRecord
+        every { attachmentClient.removeUploadData(decryptedRecord) } returns decryptedRecord
+        every { recordCryptoService.encryptRecord(decryptedRecord) } returns encryptedRecord
+        every { apiService.createRecord(ALIAS, userId, encryptedRecord) } returns Single.just(encryptedRecord)
+        every { recordCryptoService.decryptRecord(encryptedRecord, userId) } returns decryptedRecord
+        every { attachmentClient.restoreUploadData(decryptedRecord, wrappedResource, data) } returns decryptedRecord
+        every { ResourceHelper.assignResourceId(decryptedRecord) } returns decryptedRecord
+        every { SdkRecordFactory.getInstance(decryptedRecord) } returns createdRecord
 
         // When
-        try {
-            recordService.createRecord(USER_ID, doc, annotations).test().await()
-            Assert.fail("Exception expected!")
-        } catch (ex: D4LException) {
+        val subscriber = recordService.createRecord(
+                userId,
+                wrappedResource,
+                annotations
+        ).test().await()
 
-            // Then
-            Truth.assertThat(ex.javaClass).isEqualTo(DataRestrictionException.UnsupportedFileType::class.java)
-        }
-        inOrder.verify(recordService).createRecord(USER_ID, doc, annotations)
-        inOrder.verify(recordService).checkDataRestrictions(doc)
-        inOrder.verifyNoMoreInteractions()
+        val record = subscriber
+                .assertNoErrors()
+                .assertComplete()
+                .assertValueCount(1)
+                .values()[0]
+
+        // Then
+        assertSame(
+                record,
+                createdRecord
+        )
     }
 
+    // ToDo: Remove this when the switch is in the tagger
+    @Ignore //This has currently problems with mockking fhir
     @Test
-    @Throws(
-            InterruptedException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class
-    )
-    fun `Given, createData is called with data, which exceeds the file size limitation, it throws an error`() {
+    fun `Given, createRecord is called with non DataResource and a UserId, it returns a new Record`() {
         // Given
-        val invalidSizePdf = arrayOfNulls<Byte>(DATA_SIZE_MAX_BYTES + 1)
-        System.arraycopy(
-                MimeType.PDF.byteSignature()[0] as Any,
-                0,
-                invalidSizePdf,
-                0,
-                MimeType.PDF.byteSignature()[0]!!.size
-        )
-        val doc = buildDocumentReference(unboxByteArray(invalidSizePdf))
+        val wrappedResource =  mockk<WrapperContract.Resource>()
+        val rawResource = mockk<Fhir3Resource>()
+        val annotations = mockk<List<String>>()
+        val userId = "id"
+        val resourceType = "type"
 
-        val annotations = listOf<String>()
+        val data = mockk<HashMap<WrapperContract.Attachment, String?>>()
+        val decryptedRecord = mockk<NetworkRecordContract.DecryptedRecord>()
+        val tags = mockk<HashMap<String, String>>()
+        val dataKey = mockk<GCKey>()
+
+        val encryptedRecord = mockk<EncryptedRecord>()
+
+        val createdRecord = mockk<BaseRecord<Any>>()
+
+
+        every { wrappedResource.type } returns WrapperContract.Resource.TYPE.FHIR3
+        every { wrappedResource.unwrap() } returns rawResource
+
+        every { rawResource.resourceType } returns resourceType
+
+        every { attachmentService.checkDataRestrictions(wrappedResource) } returns mockk()
+        every { attachmentService.extractUploadData(wrappedResource) } returns data
+
+        every { taggingService.appendDefaultTags( resourceType, null ) } returns tags
+
+        every { cryptoService.generateGCKey() } returns Single.just(dataKey)
+
+        every {
+            anyConstructed<DecryptedRecordBuilder>().setAnnotations(annotations)
+        } returns builder
+        every { builder.build(
+                wrappedResource,
+                tags,
+                any(),//FIXME
+                dataKey,
+                ModelVersion.CURRENT
+        ) } returns decryptedRecord
+
+        every { attachmentClient.uploadData(decryptedRecord, userId) } returns decryptedRecord
+        every { attachmentClient.removeUploadData(decryptedRecord) } returns decryptedRecord
+        every { recordCryptoService.encryptRecord(decryptedRecord) } returns encryptedRecord
+        every { apiService.createRecord(ALIAS, userId, encryptedRecord) } returns Single.just(encryptedRecord)
+        every { recordCryptoService.decryptRecord(encryptedRecord, userId) } returns decryptedRecord
+        every { attachmentClient.restoreUploadData(decryptedRecord, wrappedResource, data) } returns decryptedRecord
+        every { ResourceHelper.assignResourceId(decryptedRecord) } returns decryptedRecord
+        every { SdkRecordFactory.getInstance(decryptedRecord) } returns createdRecord
 
         // When
-        try {
-            recordService.createRecord(USER_ID, doc, annotations).test().await()
-            Assert.fail("Exception expected!")
-        } catch (ex: D4LException) {
+        val subscriber = recordService.createRecord(
+                userId,
+                wrappedResource,
+                annotations
+        ).test().await()
 
-            // Then
-            Truth.assertThat(ex.javaClass).isEqualTo(DataRestrictionException.MaxDataSizeViolation::class.java)
-        }
-        inOrder.verify(recordService).createRecord(USER_ID, doc, annotations)
-        inOrder.verify(recordService).checkDataRestrictions(doc)
-        inOrder.verifyNoMoreInteractions()
+        val record = subscriber
+                .assertNoErrors()
+                .assertComplete()
+                .assertValueCount(1)
+                .values()[0]
+
+        // Then
+        assertSame(
+                record,
+                createdRecord
+        )
     }
 
+    //Todo: Add test for unhappy path
     @Test
-    @Throws(
-            InterruptedException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class
-    )
-    fun `Given, createRecords is called with multiple DomainResource and a UserId, it returns a multiple Records`() {
+    fun `Given, createRecords is called with a list of Fhir3Resources and a UserId, it returns a List of Records`() {
         // Given
-        val resources = listOf(
-                mockCarePlan as DomainResource,
-                mockCarePlan as DomainResource
-        )
-        val annotations = listOf<String>()
-        Mockito.doReturn(Single.just(mockRecord)).`when`(recordService).createRecord(USER_ID, mockCarePlan, annotations)
+        val wrappedResource =  mockk<Fhir3Resource>()
+        val resources = mutableListOf(wrappedResource)
+        val userId = "id"
+        val resourceType = "type"
+
+        val createdRecord = mockk<Record<Fhir3Resource>>()
+        every { recordService.createRecord(userId, wrappedResource, listOf()) } returns Single.just(createdRecord)
 
         // When
-        val observer = recordService.createRecords(resources, USER_ID).test().await()
+        val observer = recordService.createRecords(resources, userId).test().await()
 
         // Then
         val result = observer
@@ -333,374 +388,10 @@ class RecordServiceCreateRecordTest : RecordServiceTestBase() {
                 .assertComplete()
                 .assertValueCount(1)
                 .values()[0]
-        Truth.assertThat(result.failedOperations).isEmpty()
-        Truth.assertThat(result.successfulOperations).hasSize(2)
-        Truth.assertThat(result.successfulOperations[0].fhirResource).isEqualTo(mockCarePlan)
-        Truth.assertThat(result.successfulOperations[0].meta).isEqualTo(mockMeta)
-        Truth.assertThat(result.successfulOperations[1].fhirResource).isEqualTo(mockCarePlan)
-        Truth.assertThat(result.successfulOperations[1].meta).isEqualTo(mockMeta)
-        inOrder.verify(recordService).createRecords(resources, USER_ID)
-        inOrder.verify(recordService, Mockito.times(2)).createRecord(USER_ID, mockCarePlan, annotations)
-        inOrder.verifyNoMoreInteractions()
-    }
 
-    @Test
-    @Throws(
-            InterruptedException::class,
-            IOException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class,
-            DataValidationException.ModelVersionNotSupported::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class
-    )
-    fun `Given, createRecord is called with a DomainResource, a UserId and Annotations, it returns a new Record`() {
-        // Given
-        every {
-            anyConstructed<DecryptedRecordBuilderImpl>().setAnnotations(ANNOTATIONS)
-        } returns mockDecryptedRecordBuilder as DecryptedRecordBuilderImpl
-
-        @Suppress("UNCHECKED_CAST")
-        Mockito.`when`(
-                mockDecryptedRecordBuilder.build(
-                        mockCarePlan,
-                        mockTags,
-                        DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                        mockDataKey,
-                        ModelVersion.CURRENT
-                )
-        ).thenReturn(mockAnnotatedDecryptedFhirRecord as DecryptedFhir3Record<CarePlan>)
-        Mockito.doReturn(mockUploadData).`when`(recordService).extractUploadData(mockCarePlan)
-        Mockito.`when`(mockCarePlan.resourceType).thenReturn(CarePlan.resourceType)
-        Mockito.`when`(mockTaggingService.appendDefaultTags(
-                ArgumentMatchers.eq(CarePlan.resourceType),
-                ArgumentMatchers.any<HashMap<String, String>>())
-        ).thenReturn(mockTags)
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockDataKey))
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord).`when`(recordService)
-                ._uploadData(
-                        mockAnnotatedDecryptedFhirRecord,
-                        USER_ID
-                )
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord).`when`(recordService)
-                .removeUploadData(mockAnnotatedDecryptedFhirRecord)
-
-        Mockito.doReturn(mockAnnotatedEncryptedRecord)
-                .`when`(recordService).encryptRecord(
-                        mockAnnotatedDecryptedFhirRecord
-                )
-        Mockito.`when`(mockApiService.createRecord(ALIAS, USER_ID, mockAnnotatedEncryptedRecord))
-                .thenReturn(Single.just(mockAnnotatedEncryptedRecord))
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord)
-                .`when`(recordService)
-                .decryptRecord<DomainResource>(mockAnnotatedEncryptedRecord, USER_ID)
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord)
-                .`when`(recordService).restoreUploadData(
-                        mockAnnotatedDecryptedFhirRecord,
-                        mockCarePlan,
-                        mockUploadData
-                )
-        @Suppress("UNCHECKED_CAST")
-        every { SdkRecordFactory.getInstance(mockAnnotatedDecryptedFhirRecord) } returns mockRecord as BaseRecord<DomainResource>
-
-        // When
-        val subscriber = recordService.createRecord(USER_ID, mockCarePlan, ANNOTATIONS).test().await()
-
-        // Then
-        val record = subscriber
-                .assertNoErrors()
-                .assertComplete()
-                .assertValueCount(1)
-                .values()[0]
-
-        Truth.assertThat(record).isSameInstanceAs(mockRecord)
-
-        inOrder.verify(mockTaggingService).appendDefaultTags(
-                ArgumentMatchers.anyString(),
-                ArgumentMatchers.any<HashMap<String, String>>()
+        assertSame(
+                result.successfulOperations[0],
+                createdRecord
         )
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockDecryptedRecordBuilder).build(
-                mockCarePlan,
-                mockTags,
-                DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                mockDataKey,
-                ModelVersion.CURRENT
-        )
-        inOrder.verify(recordService)._uploadData(
-                mockAnnotatedDecryptedFhirRecord,
-                USER_ID
-        )
-        inOrder.verify(recordService).removeUploadData(mockAnnotatedDecryptedFhirRecord)
-
-        inOrder.verify(recordService).encryptRecord(
-                mockAnnotatedDecryptedFhirRecord
-        )
-        inOrder.verify(mockApiService).createRecord(ALIAS, USER_ID, mockAnnotatedEncryptedRecord)
-        inOrder.verify(recordService).decryptRecord<DomainResource>(mockAnnotatedEncryptedRecord, USER_ID)
-        inOrder.verify(recordService).restoreUploadData(
-                mockAnnotatedDecryptedFhirRecord,
-                mockCarePlan,
-                mockUploadData
-        )
-        inOrder.verify(recordService).assignResourceId(mockAnnotatedDecryptedFhirRecord)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(
-            InterruptedException::class,
-            IOException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class,
-            DataValidationException.ModelVersionNotSupported::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class
-    )
-    fun `Given, createRecords is called a DomainResource and a UserId, Annotations and without attachment, it returns a new Record`() {
-        // Given
-        every {
-            anyConstructed<DecryptedRecordBuilderImpl>().setAnnotations(ANNOTATIONS)
-        } returns mockDecryptedRecordBuilder as DecryptedRecordBuilderImpl
-
-        @Suppress("UNCHECKED_CAST")
-        Mockito.`when`(
-                mockDecryptedRecordBuilder.build(
-                        mockCarePlan,
-                        mockTags,
-                        DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                        mockDataKey,
-                        ModelVersion.CURRENT
-                )
-        ).thenReturn(mockAnnotatedDecryptedFhirRecord as DecryptedFhir3Record<CarePlan>)
-        Mockito.doReturn(null).`when`(recordService).extractUploadData(mockCarePlan)
-        Mockito.`when`(mockCarePlan.resourceType).thenReturn(CarePlan.resourceType)
-        Mockito.`when`(
-                mockTaggingService.appendDefaultTags(
-                        ArgumentMatchers.eq(CarePlan.resourceType),
-                        ArgumentMatchers.any<HashMap<String, String>>()
-                )
-        ).thenReturn(mockTags)
-        Mockito.`when`(mockCryptoService.generateGCKey())
-                .thenReturn(Single.just(mockDataKey))
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord)
-                .`when`(recordService)
-                ._uploadData(
-                        mockAnnotatedDecryptedFhirRecord,
-                        USER_ID
-                )
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord)
-                .`when`(recordService)
-                .removeUploadData(mockAnnotatedDecryptedFhirRecord)
-
-        Mockito.doReturn(mockAnnotatedEncryptedRecord).`when`(recordService)
-                .encryptRecord(mockAnnotatedDecryptedFhirRecord)
-        Mockito.`when`(mockApiService.createRecord(ALIAS, USER_ID, mockAnnotatedEncryptedRecord))
-                .thenReturn(Single.just(mockAnnotatedEncryptedRecord))
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord)
-                .`when`(recordService)
-                .decryptRecord<DomainResource>(
-                        mockAnnotatedEncryptedRecord,
-                        USER_ID
-                )
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord)
-                .`when`(recordService)
-                .restoreUploadData(
-                        mockAnnotatedDecryptedFhirRecord,
-                        mockCarePlan,
-                        mockUploadData
-                )
-
-        Mockito.doReturn(mockAnnotatedDecryptedFhirRecord)
-                .`when`(recordService)
-                ._uploadData(
-                        mockAnnotatedDecryptedFhirRecord,
-                        USER_ID
-                )
-        @Suppress("UNCHECKED_CAST")
-        every { SdkRecordFactory.getInstance(mockAnnotatedDecryptedFhirRecord) } returns mockRecord as BaseRecord<DomainResource>
-
-        // When
-        val subscriber = recordService.createRecord(USER_ID, mockCarePlan, ANNOTATIONS).test().await()
-
-        // Then
-        val record = subscriber
-                .assertNoErrors()
-                .assertComplete()
-                .assertValueCount(1)
-                .values()[0]
-
-        Truth.assertThat(record).isSameInstanceAs(mockRecord)
-
-        inOrder.verify(mockTaggingService).appendDefaultTags(
-                ArgumentMatchers.anyString(),
-                ArgumentMatchers.any<HashMap<String, String>>()
-        )
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockDecryptedRecordBuilder).build(
-                mockCarePlan,
-                mockTags,
-                DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                mockDataKey,
-                ModelVersion.CURRENT
-        )
-        inOrder.verify(recordService)._uploadData(
-                mockAnnotatedDecryptedFhirRecord,
-                USER_ID
-        )
-        inOrder.verify(recordService).removeUploadData(
-                mockAnnotatedDecryptedFhirRecord
-        )
-
-        inOrder.verify(recordService).encryptRecord(
-                mockAnnotatedDecryptedFhirRecord
-        )
-        inOrder.verify(mockApiService).createRecord(ALIAS, USER_ID, mockAnnotatedEncryptedRecord)
-        inOrder.verify(recordService).decryptRecord<DomainResource>(mockAnnotatedEncryptedRecord, USER_ID)
-        inOrder.verify(recordService).restoreUploadData(
-                mockAnnotatedDecryptedFhirRecord,
-                mockCarePlan,
-                null
-        )
-        inOrder.verify(recordService).assignResourceId(mockAnnotatedDecryptedFhirRecord)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(
-            InterruptedException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class
-    )
-    fun `Given, createRecord is called with a unsupported Data, a UserId and Annotations, it throws an error`() {
-        // Given
-        val invalidData = byteArrayOf(0x00)
-        val doc = buildDocumentReference(invalidData)
-
-        // When
-        try {
-            recordService.createRecord(USER_ID, doc, ANNOTATIONS).test().await()
-            Assert.fail("Exception expected!")
-        } catch (ex: D4LException) {
-
-            // Then
-            Truth.assertThat(ex.javaClass).isEqualTo(DataRestrictionException.UnsupportedFileType::class.java)
-        }
-        inOrder.verify(recordService).createRecord(USER_ID, doc, ANNOTATIONS)
-        inOrder.verify(recordService).checkDataRestrictions(doc)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Throws(
-            InterruptedException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class
-    )
-    fun `Given, createRecord is called with data, which exceeds the file size limitation, a UserId and Annotations, it throws an error`() {
-        // Given
-        val invalidSizePdf = arrayOfNulls<Byte>(DATA_SIZE_MAX_BYTES + 1)
-        System.arraycopy(
-                MimeType.PDF.byteSignature()[0] as Any,
-                0,
-                invalidSizePdf,
-                0,
-                MimeType.PDF.byteSignature()[0]!!.size
-        )
-        val doc = buildDocumentReference(unboxByteArray(invalidSizePdf))
-
-        // When
-        try {
-            recordService.createRecord(USER_ID, doc, ANNOTATIONS).test().await()
-            Assert.fail("Exception expected!")
-        } catch (ex: D4LException) {
-
-            // Then
-            Truth.assertThat(ex.javaClass).isEqualTo(DataRestrictionException.MaxDataSizeViolation::class.java)
-        }
-        inOrder.verify(recordService).createRecord(USER_ID, doc, ANNOTATIONS)
-        inOrder.verify(recordService).checkDataRestrictions(doc)
-        inOrder.verifyNoMoreInteractions()
-    }
-
-    @Test
-    @Ignore
-    @Throws(
-            InterruptedException::class,
-            IOException::class,
-            DataRestrictionException.UnsupportedFileType::class,
-            DataRestrictionException.MaxDataSizeViolation::class,
-            DataValidationException.ModelVersionNotSupported::class,
-            DataValidationException.ExpectedFieldViolation::class,
-            DataValidationException.IdUsageViolation::class,
-            DataValidationException.InvalidAttachmentPayloadHash::class
-    )
-    fun `Given createRecord is called with a Byte resource, a UserId and Annotations, it returns a new DataRecord`() {
-        // Given
-        every {
-            anyConstructed<DecryptedRecordBuilderImpl>().setAnnotations(ANNOTATIONS)
-        } returns mockDecryptedRecordBuilder as DecryptedRecordBuilderImpl
-
-        Mockito.`when`(mockTaggingService.appendDefaultTags(null, null))
-                .thenReturn(mockTags)
-        Mockito.`when`(mockCryptoService.generateGCKey()).thenReturn(Single.just(mockDataKey))
-        @Suppress("UNCHECKED_CAST")
-        Mockito.`when`(
-                mockDecryptedRecordBuilder.build(
-                        mockDataResource.value,
-                        mockTags,
-                        DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                        mockDataKey,
-                        ModelVersion.CURRENT
-                )
-        ).thenReturn(mockDecryptedDataRecord)
-        Mockito.doReturn(mockEncryptedRecord)
-                .`when`(recordService)
-                .encryptRecord(mockDecryptedDataRecord)
-        Mockito.`when`(mockApiService.createRecord(ALIAS, USER_ID, mockEncryptedRecord))
-                .thenReturn(Single.just(mockEncryptedRecord))
-        Mockito.doReturn(mockDecryptedDataRecord)
-                .`when`(recordService)
-                .decryptRecord<ByteArray>(mockEncryptedRecord, USER_ID)
-        Mockito.doReturn(mockDecryptedDataRecord)
-                .`when`(recordService)
-                .assignResourceId(mockDecryptedDataRecord)
-        @Suppress("UNCHECKED_CAST")
-        every { SdkRecordFactory.getInstance(mockDecryptedDataRecord) } returns mockDataRecord
-
-        // When
-        val subscriber = recordService.createRecord(
-                USER_ID,
-                mockDataResource,
-                ANNOTATIONS
-        ).test().await()
-
-        // Then
-        val record = subscriber
-                .assertNoErrors()
-                .assertComplete()
-                .assertValueCount(1)
-                .values()[0]
-        Truth.assertThat(record).isSameInstanceAs(mockDataRecord)
-
-        inOrder.verify(mockTaggingService).appendDefaultTags(null, null)
-        inOrder.verify(mockCryptoService).generateGCKey()
-        inOrder.verify(mockDecryptedRecordBuilder).build(
-                mockDataResource.value,
-                mockTags,
-                DATE_FORMATTER.format(LocalDate.now(UTC_ZONE_ID)),
-                mockDataKey,
-                ModelVersion.CURRENT
-        )
-        inOrder.verify(recordService)
-                .encryptRecord(mockDecryptedDataRecord)//mockDecryptedDataRecord)
-        inOrder.verify(mockApiService).createRecord(ALIAS, USER_ID, mockEncryptedRecord)
-        inOrder.verify(recordService).decryptRecord<ByteArray>(mockEncryptedRecord, USER_ID)
-        inOrder.verify(recordService).assignResourceId(mockDecryptedDataRecord)
-        inOrder.verifyNoMoreInteractions()
     }
 }
-
- */
