+++ sdk-core/src/test/java/care/data4life/sdk/dropme/RecordServiceAdditionalResourceTypeTest.kt
@@ -13,7 +13,7 @@
  * applications and/or if you’d like to contribute to the development of the SDK, please
  * contact D4L by email to help@data4life.care.
  */
-package care.data4life.sdk
+package care.data4life.sdk.dropme
 
 /*
 class RecordServiceAdditionalResourceTypeTest : RecordServiceTestBase() {
