+++ sdk-core/src/test/java/care/data4life/sdk/dropme/RecordServiceTestBase.kt
@@ -14,7 +14,7 @@
  * contact D4L by email to help@data4life.care.
  */
 
-package care.data4life.sdk
+package care.data4life.sdk.dropme
 
 /*
 abstract class RecordServiceTestBase {
